Rename package for a better understanding
