+++ go/src/koding/kites/kloud/provider/softlayer/build.go
@@ -7,7 +7,7 @@ import (
 	"koding/kites/kloud/api/sl"
 	"koding/kites/kloud/contexthelper/publickeys"
 	"koding/kites/kloud/machinestate"
-	"koding/kites/kloud/scripts/softlayer/userdata"
+	"koding/kites/kloud/scripts/provisionklient/userdata"
 	"strings"
 	"time"
 
@@ -22,8 +22,8 @@ import (
 )
 
 var (
-	// Go binary source code residues at go/src/koding/kites/kloud/scripts/softlayer
-	PostInstallScriptUri = "https://s3.amazonaws.com/kodingdev-softlayer/softlayer"
+	// Go binary source code residues at go/src/koding/kites/kloud/scripts/provisionklient
+	PostInstallScriptUri = "https://s3.amazonaws.com/kodingdev-provision/provisionklient"
 
 	// Only lookup images that have this tag
 	DefaultTemplateTag = "koding-stable"
