+++ go/src/koding/kites/kloud/scripts/provisionklient/userdata/userdata.go
